+++ src/components/CheerList.js
@@ -1,0 +1,21 @@
+import React, { Fragment } from "react";
+
+import Cheer, { COLORS } from "./Cheer";
+import CheerButton from "./CheerButton";
+
+const colorKeys = Object.keys(COLORS);
+
+const CheerList = ({ cheered, cheers, name, onCheerToggle }) => (
+  <Fragment>
+    <CheerButton cheered={cheered} name={name} onCheerToggle={onCheerToggle} />
+    {cheers.map(({ name: cheerName }, index) => (
+      <Cheer
+        key={`${index}-${name}`} // eslint-disable-line
+        color={colorKeys[((cheered ? 1 : 0) + index) % colorKeys.length]}
+        name={cheerName}
+      />
+    ))}
+  </Fragment>
+);
+
+export default CheerList;