+++ src/index.js
@@ -1,11 +1,19 @@
 import "./styles/app.scss";
 
 export { default as Badge } from "./components/Badge";
+export { default as BooleanField } from "./components/BooleanField";
 export { default as Button } from "./components/Button";
+export { default as CentsField } from "./components/CentsField";
 export { default as Checklist } from "./components/Checklist";
 export { default as Checkmark } from "./components/Checkmark";
+export { default as Cheer } from "./components/Cheer";
+export { default as CheerButton } from "./components/CheerButton";
+export { default as CheerList } from "./components/CheerList";
 export { default as Circles } from "./components/Circles";
+export { EmailField, NumberField, PasswordField, StringField } from "./components/FormFields";
 export { default as FeedItem } from "./components/FeedItem";
+export { default as FileField } from "./components/FileField";
+export { default as Form } from "./components/Form";
 export { default as Hamburger } from "./components/Hamburger";
 export { default as Loader } from "./components/Loader";
 export { default as Icon } from "./components/Icon";
@@ -14,15 +22,10 @@ export { default as Nav } from "./components/Nav";
 export { default as Pagination } from "./components/Pagination";
 export { default as Panel } from "./components/Panel";
 export { default as Spinner } from "./components/Spinner";
+export { default as SubmitButton } from "./components/SubmitButton";
 export { default as Subnav } from "./components/Subnav";
 export { default as Success } from "./components/Success";
 export { default as Tag } from "./components/Tag";
 export { default as Thumbnail } from "./components/Thumbnail";
+export { default as Tooltip } from "./components/Tooltip";
 export { default as Warning } from "./components/Warning";
-
-export { default as BooleanField } from "./components/BooleanField";
-export { default as CentsField } from "./components/CentsField";
-export { default as FileField } from "./components/FileField";
-export { default as Form } from "./components/Form";
-export { EmailField, NumberField, PasswordField, StringField } from "./components/FormFields";
-export { default as SubmitButton } from "./components/SubmitButton";