+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- The `Cheer` component.
+- The `CheerButton` component.
+- The `Tooltip` component.
+- The `CheerList` component.
 
 ## [3.4.0] - 2018-09-13
 ### Changed